requirements.txt
scikit-learn
pygeoip
python-whois
tldextract
pydot
pandas